style: rename view params editor components
