+++ apps/client/src/common/components/view-params-editor/ViewParamsEditor.tsx
@@ -12,16 +12,16 @@ import {
   useDisclosure,
 } from '@chakra-ui/react';
 
-import EditFormInput from './EditFormInput';
+import ParamInput from './ParamInput';
 import { ParamField } from './types';
 
-import style from './EditFormDrawer.module.scss';
+import style from './ViewParamsEditor.module.scss';
 
 interface EditFormDrawerProps {
   paramFields: ParamField[];
 }
 
-export default function EditFormDrawer({ paramFields }: EditFormDrawerProps) {
+export default function ViewParamsEditor({ paramFields }: EditFormDrawerProps) {
   const [searchParams, setSearchParams] = useSearchParams();
   const { isOpen, onClose, onOpen } = useDisclosure();
 
@@ -77,7 +77,7 @@ export default function EditFormDrawer({ paramFields }: EditFormDrawerProps) {
                 <label className={style.label}>
                   <span className={style.title}>{field.title}</span>
                   <span className={style.description}>{field.description}</span>
-                  <EditFormInput key={field.title} paramField={field} />
+                  <ParamInput key={field.title} paramField={field} />
                 </label>
               </div>
             ))}
